add session tracking

+++ ui/src/ReadingSessions.js
@@ -1,9 +1,11 @@
 import { Component } from 'preact';
 import { html } from 'htm/preact';
-import { route } from 'preact-router';
 
 class ReadingSession extends Component {
 	delete(ts, refresh) {
+		if (!confirm("Are you sure you want to delete this session?")) {
+			return;
+		}
 		fetch("/api/reading_sessions/" + ts, {
 			method: "DELETE"
 		}).then(() => {
@@ -13,8 +15,10 @@ class ReadingSession extends Component {
 
 	render({ session, refresh }) {
 		return html`
-			<span><strong>${new Date(session.timestamp*1000).toLocaleString()}:</strong> ${session.duration} min
-			- <a onclick=${(function() { this.delete(session.timestamp, refresh) }).bind(this)}>Delete</a></span>
+			<span style="line-height: 40px">
+				<a class='rfa-button rfa-button-small' onclick=${(function() { this.delete(session.timestamp, refresh) }).bind(this)}>Delete</a>
+				<strong>${new Date(session.timestamp*1000).toLocaleString()}:</strong> ${session.duration} min
+			</span>
 		`
 	}
 }
@@ -56,6 +60,82 @@ class AddReadingSession extends Component {
 	}
 }
 
+class RecordReadingSession extends Component {
+	constructor() {
+	  super();
+	  // set initial time:
+	  this.state = {
+		duration: 0,
+		start: null,
+		now: null,
+	  };
+	}
+
+	start() {
+		this.setState({
+		  start: new Date(),
+		now: new Date(),
+	  })
+	  this.timer = setInterval(() => {
+		this.setState({
+				  now: new Date(),
+		});
+	  }, 1000);
+	}
+
+	stop() {
+		this.setState({
+		  duration: this.state.duration + (this.state.now - this.state.start),
+		start: null,
+		now: null,
+	  })
+		clearInterval(this.timer);
+	}
+
+	reset() {
+		this.setState({ duration: 0, start: null, now: null, })
+	}
+
+	submit(refresh) {
+		return (function(e) {
+			e.preventDefault();
+			const totalMinutes = Math.floor(totalSeconds / 60);
+			fetch("/api/reading_sessions", {
+				method: "POST",
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				body: JSON.stringify({ "duration": totalMinutes }),
+			}).then(() => {
+				refresh()
+			})
+			this.setState({ duration: 0 })
+			e.target.reset()
+		}.bind(this))
+	}
+
+	componentWillUnmount() {
+		clearInterval(this.timer);
+	}
+  
+	render({refresh}, state) {
+		let totalDuration = state.duration;
+		if (state.now && state.start) {
+			totalDuration += (state.now - state.start);
+		}
+		const totalSeconds = Math.floor(totalDuration/1000);
+		const totalMinutes = Math.floor(totalSeconds / 60);
+		const seconds = totalSeconds%60;
+		return html `
+			<div>Duration: <strong>${totalMinutes < 10 ? '0'+totalMinutes : totalMinutes}:${seconds < 10 ? '0'+seconds : seconds}</strong></div>
+			<a class='rfa-button' onclick=${this.start.bind(this)}>Start</a>
+			<a class='rfa-button' onclick=${this.stop.bind(this)}>Stop</a>
+			<a class='rfa-button' onclick=${this.reset.bind(this)}>Reset</a>
+			<a class='rfa-button' onclick=${this.submit(refresh)}>Submit</a>
+		`;
+	}
+  }
+
 class ReadingSessions extends Component {
 	constructor() {
 		super()
@@ -96,12 +176,23 @@ class ReadingSessions extends Component {
 
 		return html`
 			<h3>Sessions</h3>
-			${this.state.sessions.map(session => html`
-				<li><${ReadingSession} session=${session} refresh=${this.refreshList.bind(this)} /></li>
-			`)}
+			<div>
+				<h4>Record a session</h4>
+				<p>Start recording a session.</p>
+				<${RecordReadingSession} refresh=${this.refreshList.bind(this)} />
+			</div>
 
-			<h4>Add a session</h4>
-			<${AddReadingSession} refresh=${this.refreshList.bind(this)} />
+			<div>
+				<h4>Previous sessions</h4>
+				${this.state.sessions.map(session => html`
+					<li><${ReadingSession} session=${session} refresh=${this.refreshList.bind(this)} /></li>
+				`)}
+			</div>
+
+			<div>
+				<h4>Add a session</h4>
+				<${AddReadingSession} refresh=${this.refreshList.bind(this)} />
+			</div>
 		`
 	}
 }
